alerts created

+++ src/components/Products.js
@@ -1,5 +1,8 @@
 import React, { useState } from "react";
 import { Button, Form } from "react-bootstrap";
+
+import swal from "sweetalert";
+
 import "./index.css";
 
 import { BrowserRouter as Router, useParams } from "react-router-dom";
@@ -40,21 +43,39 @@ const Products = (props) => {
         setCategory("");
         const newProducts = [...products, data];
         setProducts(newProducts);
+        swal("Creating your product!", "success");
       }
     } catch (error) {
       console.error(error);
     }
   };
 
+  // Swal message for delete works? state issue
   const handleProductsDelete = async (id) => {
     try {
-      const data = await deleteProduct(id, token);
-      if (data) {
-        // const newProducts;
-        console.log("products:", products);
-        // console.log("newProducts:", newProducts);
-        setProducts(products.filter((product) => data.id !== product.id));
-      }
+      swal({
+        title: "Are you sure?",
+        text: "Once deleted, you will not be able to recover this product!",
+        icon: "warning",
+        buttons: true,
+        dangerMode: true,
+      }).then((willDelete) => {
+        if (willDelete) {
+          // handleProductsDelete();
+          const data = deleteProduct(id, token);
+          if (data) {
+            // const newProducts;
+            // console.log("products:", products);
+            // console.log("newProducts:", newProducts);
+            setProducts(products.filter((product) => data.id !== product.id));
+          }
+          swal("Your product has been deleted!", {
+            icon: "success",
+          });
+        } else {
+          swal("Your imaginary file is safe!");
+        }
+      });
     } catch (error) {
       console.error(error);
     }
@@ -147,6 +168,7 @@ const Products = (props) => {
             return (
               <React.Fragment key={product.id}>
                 <SingleProduct
+                  handleConfirmDelete={handleProductsDelete}
                   handleProductsDelete={handleProductsDelete}
                   products={products}
                   setProducts={setProducts}
